fix: move password service
