+++ backend/AASTU-backend-group-7/Infrastructure/password_services/password_service.go
@@ -1,4 +1,4 @@
-package Infrastructure
+package password_services
 
 import "golang.org/x/crypto/bcrypt"
